Extract graphqlbatch and the message transformer into graphQLUtils.js
